backend/flask-api/requirements.txt
Flask
Flask-Redis
sentence-transformers
PyPDF2
faiss-cpu
openai
